-user login problem is fixed and run smoothly.

+++ CMPE312-PROJECT/Models/Entity/User.cs
@@ -12,7 +12,7 @@ namespace CMPE312_PROJECT.Models.Entity
         public String email { get; set; }
         public String salt { get; set; }
         public String passwordHash { get; set; }
-        public int isAdmin { get; set; }
+        public decimal isAdmin { get; set; }
         public String status { get; set; }
 
         public User()
